+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Cryptography/DecryptParameters.cs
@@ -8,134 +8,134 @@ using Azure.Core;
 namespace Azure.Security.KeyVault.Keys.Cryptography
 {
     /// <summary>
-    /// Options for decrypting ciphertext.
+    /// Parameters for decrypting ciphertext.
     /// </summary>
-    public class DecryptOptions : IJsonSerializable
+    public class DecryptParameters : IJsonSerializable
     {
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.Rsa15"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.Rsa15"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.Rsa15"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.Rsa15"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> is null.</exception>
-        public static DecryptOptions Rsa15Options(byte[] ciphertext) =>
-            new DecryptOptions(EncryptionAlgorithm.Rsa15, ciphertext);
+        public static DecryptParameters Rsa15Parameters(byte[] ciphertext) =>
+            new DecryptParameters(EncryptionAlgorithm.Rsa15, ciphertext);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.RsaOaep"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.RsaOaep"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.RsaOaep"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.RsaOaep"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> is null.</exception>
-        public static DecryptOptions RsaOaepOptions(byte[] ciphertext) =>
-            new DecryptOptions(EncryptionAlgorithm.RsaOaep, ciphertext);
+        public static DecryptParameters RsaOaepParameters(byte[] ciphertext) =>
+            new DecryptParameters(EncryptionAlgorithm.RsaOaep, ciphertext);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.RsaOaep256"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.RsaOaep256"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.RsaOaep256"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.RsaOaep256"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> is null.</exception>
-        public static DecryptOptions RsaOaep256Options(byte[] ciphertext) =>
-            new DecryptOptions(EncryptionAlgorithm.RsaOaep256, ciphertext);
+        public static DecryptParameters RsaOaep256Parameters(byte[] ciphertext) =>
+            new DecryptParameters(EncryptionAlgorithm.RsaOaep256, ciphertext);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A128Gcm"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A128Gcm"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector (or nonce) generated during encryption.</param>
         /// <param name="authenticationTag">The authentication tag generated during encryption.</param>
         /// <param name="additionalAuthenticationData">Optional data that is authenticated but not encrypted.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A128Gcm"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A128Gcm"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/>, <paramref name="iv"/>, or <paramref name="authenticationTag"/> is null.</exception>
-        public static DecryptOptions A128GcmOptions(byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticationData = null) =>
-            new DecryptOptions(EncryptionAlgorithm.A128Gcm, ciphertext, iv, authenticationTag, additionalAuthenticationData);
+        public static DecryptParameters A128GcmParameters(byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticationData = null) =>
+            new DecryptParameters(EncryptionAlgorithm.A128Gcm, ciphertext, iv, authenticationTag, additionalAuthenticationData);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A192Gcm"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A192Gcm"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector (or nonce) generated during encryption.</param>
         /// <param name="authenticationTag">The authentication tag generated during encryption.</param>
         /// <param name="additionalAuthenticationData">Optional data that is authenticated but not encrypted.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A192Gcm"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A192Gcm"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/>, <paramref name="iv"/>, or <paramref name="authenticationTag"/> is null.</exception>
-        public static DecryptOptions A192GcmOptions(byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticationData = null) =>
-            new DecryptOptions(EncryptionAlgorithm.A192Gcm, ciphertext, iv, authenticationTag, additionalAuthenticationData);
+        public static DecryptParameters A192GcmParameters(byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticationData = null) =>
+            new DecryptParameters(EncryptionAlgorithm.A192Gcm, ciphertext, iv, authenticationTag, additionalAuthenticationData);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A256Gcm"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A256Gcm"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector (or nonce) generated during encryption.</param>
         /// <param name="authenticationTag">The authentication tag generated during encryption.</param>
         /// <param name="additionalAuthenticationData">Optional data that is authenticated but not encrypted.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A256Gcm"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A256Gcm"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/>, <paramref name="iv"/>, or <paramref name="authenticationTag"/> is null.</exception>
-        public static DecryptOptions A256GcmOptions(byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticationData = null) =>
-            new DecryptOptions(EncryptionAlgorithm.A256Gcm, ciphertext, iv, authenticationTag, additionalAuthenticationData);
+        public static DecryptParameters A256GcmParameters(byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticationData = null) =>
+            new DecryptParameters(EncryptionAlgorithm.A256Gcm, ciphertext, iv, authenticationTag, additionalAuthenticationData);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A128Cbc"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A128Cbc"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector used during encryption.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A128Cbc"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A128Cbc"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> or <paramref name="iv"/> is null.</exception>
-        public static DecryptOptions A128CbcOptions(byte[] ciphertext, byte[] iv) =>
-            new DecryptOptions(EncryptionAlgorithm.A128Cbc, ciphertext, iv);
+        public static DecryptParameters A128CbcParameters(byte[] ciphertext, byte[] iv) =>
+            new DecryptParameters(EncryptionAlgorithm.A128Cbc, ciphertext, iv);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A192Cbc"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A192Cbc"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector used during encryption.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A192Cbc"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A192Cbc"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> or <paramref name="iv"/> is null.</exception>
-        public static DecryptOptions A192CbcOptions(byte[] ciphertext, byte[] iv) =>
-            new DecryptOptions(EncryptionAlgorithm.A192Cbc, ciphertext, iv);
+        public static DecryptParameters A192CbcParameters(byte[] ciphertext, byte[] iv) =>
+            new DecryptParameters(EncryptionAlgorithm.A192Cbc, ciphertext, iv);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A256Cbc"/> encryption algorithm.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A256Cbc"/> encryption algorithm.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector used during encryption.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A256Cbc"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A256Cbc"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> or <paramref name="iv"/> is null.</exception>
-        public static DecryptOptions A256CbcOptions(byte[] ciphertext, byte[] iv) =>
-            new DecryptOptions(EncryptionAlgorithm.A256Cbc, ciphertext, iv);
+        public static DecryptParameters A256CbcParameters(byte[] ciphertext, byte[] iv) =>
+            new DecryptParameters(EncryptionAlgorithm.A256Cbc, ciphertext, iv);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A128CbcPad"/> encryption algorithm with PKCS#7 padding.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A128CbcPad"/> encryption algorithm with PKCS#7 padding.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector used during encryption.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A128CbcPad"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A128CbcPad"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> or <paramref name="iv"/> is null.</exception>
-        public static DecryptOptions A128CbcPadOptions(byte[] ciphertext, byte[] iv) =>
-            new DecryptOptions(EncryptionAlgorithm.A128CbcPad, ciphertext, iv);
+        public static DecryptParameters A128CbcPadParameters(byte[] ciphertext, byte[] iv) =>
+            new DecryptParameters(EncryptionAlgorithm.A128CbcPad, ciphertext, iv);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A192CbcPad"/> encryption algorithm with PKCS#7 padding.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A192CbcPad"/> encryption algorithm with PKCS#7 padding.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector used during encryption.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A192CbcPad"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A192CbcPad"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> or <paramref name="iv"/> is null.</exception>
-        public static DecryptOptions A192CbcPadOptions(byte[] ciphertext, byte[] iv) =>
-            new DecryptOptions(EncryptionAlgorithm.A192CbcPad, ciphertext, iv);
+        public static DecryptParameters A192CbcPadParameters(byte[] ciphertext, byte[] iv) =>
+            new DecryptParameters(EncryptionAlgorithm.A192CbcPad, ciphertext, iv);
 
         /// <summary>
-        /// Creates an instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A256CbcPad"/> encryption algorithm with PKCS#7 padding.
+        /// Creates an instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A256CbcPad"/> encryption algorithm with PKCS#7 padding.
         /// </summary>
         /// <param name="ciphertext">The ciphertext to decrypt.</param>
         /// <param name="iv">The initialization vector used during encryption.</param>
-        /// <returns>An instance of the <see cref="DecryptOptions"/> class for the <see cref="EncryptionAlgorithm.A256CbcPad"/> encryption algorithm.</returns>
+        /// <returns>An instance of the <see cref="DecryptParameters"/> class for the <see cref="EncryptionAlgorithm.A256CbcPad"/> encryption algorithm.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="ciphertext"/> or <paramref name="iv"/> is null.</exception>
-        public static DecryptOptions A256CbcPadOptions(byte[] ciphertext, byte[] iv) =>
-            new DecryptOptions(EncryptionAlgorithm.A256CbcPad, ciphertext, iv);
+        public static DecryptParameters A256CbcPadParameters(byte[] ciphertext, byte[] iv) =>
+            new DecryptParameters(EncryptionAlgorithm.A256CbcPad, ciphertext, iv);
 
-        internal DecryptOptions(EncryptionAlgorithm algorithm, byte[] ciphertext)
+        internal DecryptParameters(EncryptionAlgorithm algorithm, byte[] ciphertext)
         {
             Argument.AssertNotNull(ciphertext, nameof(ciphertext));
 
@@ -143,7 +143,7 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
             Ciphertext = ciphertext;
         }
 
-        internal DecryptOptions(EncryptionAlgorithm algorithm, byte[] ciphertext, byte[] iv)
+        internal DecryptParameters(EncryptionAlgorithm algorithm, byte[] ciphertext, byte[] iv)
         {
             Argument.AssertNotNull(ciphertext, nameof(ciphertext));
             Argument.AssertNotNull(iv, nameof(iv));
@@ -153,7 +153,7 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
             Iv = iv;
         }
 
-        internal DecryptOptions(EncryptionAlgorithm algorithm, byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticatedData)
+        internal DecryptParameters(EncryptionAlgorithm algorithm, byte[] ciphertext, byte[] iv, byte[] authenticationTag, byte[] additionalAuthenticatedData)
         {
             Argument.AssertNotNull(ciphertext, nameof(ciphertext));
             Argument.AssertNotNull(iv, nameof(iv));