+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Cryptography/ICryptographyProvider.cs
@@ -12,13 +12,13 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
 
         bool SupportsOperation(KeyOperation operation);
 
-        Task<EncryptResult> EncryptAsync(EncryptOptions options, CancellationToken cancellationToken = default);
+        Task<EncryptResult> EncryptAsync(EncryptParameters parameters, CancellationToken cancellationToken = default);
 
-        EncryptResult Encrypt(EncryptOptions options, CancellationToken cancellationToken = default);
+        EncryptResult Encrypt(EncryptParameters parameters, CancellationToken cancellationToken = default);
 
-        Task<DecryptResult> DecryptAsync(DecryptOptions options, CancellationToken cancellationToken = default);
+        Task<DecryptResult> DecryptAsync(DecryptParameters parameters, CancellationToken cancellationToken = default);
 
-        DecryptResult Decrypt(DecryptOptions options, CancellationToken cancellationToken = default);
+        DecryptResult Decrypt(DecryptParameters parameters, CancellationToken cancellationToken = default);
 
         Task<WrapResult> WrapKeyAsync(KeyWrapAlgorithm algorithm, byte[] key, CancellationToken cancellationToken = default);